Attempt to handle LIMIT in Impala
Impala is teh dum when it comes to applying limits.  If the limit is
applied to a subquery that contains an order by statement, that order by
statement is ignored.  No other RDBMS does this, but it means that we
have to manually append limit statements onto the end of the SQL query
instead of used Dataset#limit